+++ app/controllers/users/webauthn_setup_controller.rb
@@ -8,7 +8,6 @@ module Users
     include ReauthenticationRequiredConcern
     include ThreatMetrixHelper
     include ThreatMetrixConcern
-    include WebauthnSetupFormConcern
 
     before_action :authenticate_user!
     before_action :confirm_user_authenticated_for_2fa_setup
@@ -279,6 +278,47 @@ module Users
         :platform_authenticator,
         :transports,
       ).merge(protocol: request.protocol)
+    end
+
+    def prepare_webauthn_setup_form(
+      platform_authenticator:,
+      auto_trigger:,
+      need_to_set_up_additional_mfa:,
+      passkey_upsell: false
+    )
+      save_challenge_in_session
+      @exclude_credentials = exclude_credentials
+      @platform_authenticator = platform_authenticator
+      @auto_trigger = auto_trigger
+      @passkey_upsell = passkey_upsell
+      @presenter = build_webauthn_setup_presenter(
+        platform_authenticator: @platform_authenticator,
+        passkey_upsell: @passkey_upsell,
+      )
+      @need_to_set_up_additional_mfa = need_to_set_up_additional_mfa
+      @account_creation_threatmetrix ||= account_creation_threatmetrix_variables
+      user_session[:webauthn_setup_started_at] = Time.zone.now.to_f if platform_authenticator
+    end
+
+    def save_challenge_in_session
+      credential_creation_options = WebAuthn::Credential.options_for_create(user: current_user)
+      user_session[:webauthn_challenge] = credential_creation_options.challenge.bytes.to_a
+    end
+
+    def exclude_credentials
+      current_user.webauthn_configurations.map(&:credential_id)
+    end
+
+    def build_webauthn_setup_presenter(platform_authenticator:, passkey_upsell: false)
+      WebauthnSetupPresenter.new(
+        current_user: current_user,
+        user_fully_authenticated: user_fully_authenticated?,
+        user_opted_remember_device_cookie: user_opted_remember_device_cookie,
+        remember_device_default: remember_device_default,
+        platform_authenticator: platform_authenticator,
+        passkey_upsell: passkey_upsell,
+        url_options:,
+      )
     end
   end
 end
